+++ rust/bridge/shared/src/net/chat.rs
@@ -30,7 +30,7 @@ use crate::*;
 bridge_handle_fns!(HttpRequest, clone = false);
 bridge_handle_fns!(UnauthenticatedChatConnection, clone = false);
 bridge_handle_fns!(AuthenticatedChatConnection, clone = false);
-bridge_handle_fns!(ProvisioningChatConnection, clone = false, jni = false);
+bridge_handle_fns!(ProvisioningChatConnection, clone = false);
 
 #[bridge_fn(ffi = false)]
 fn HttpRequest_new(
@@ -254,14 +254,14 @@ fn ServerMessageAck_SendStatus(
     sender(status.into_inner().into())
 }
 
-#[bridge_io(TokioAsyncContext, jni = false)]
+#[bridge_io(TokioAsyncContext)]
 async fn ProvisioningChatConnection_connect(
     connection_manager: &ConnectionManager,
 ) -> Result<ProvisioningChatConnection, ConnectError> {
     ProvisioningChatConnection::connect(connection_manager).await
 }
 
-#[bridge_fn(jni = false)]
+#[bridge_fn]
 fn ProvisioningChatConnection_init_listener(
     chat: &ProvisioningChatConnection,
     listener: Box<dyn ProvisioningListener>,
@@ -274,7 +274,7 @@ fn ProvisioningChatConnection_info(chat: &ProvisioningChatConnection) -> ChatCon
     chat.info()
 }
 
-#[bridge_io(TokioAsyncContext, jni = false)]
+#[bridge_io(TokioAsyncContext)]
 async fn ProvisioningChatConnection_disconnect(chat: &ProvisioningChatConnection) {
     chat.disconnect().await
 }